Fix module name clashing and add docs
