+++ .dockerignore
@@ -6,4 +6,4 @@ npm-debug.log
 /index.js
 /index.mjs
 /index.d.ts
-/utils.js
+/sdk.js
