+++ .gitignore
@@ -2,7 +2,7 @@
 /index.js
 /index.mjs
 /index.d.ts
-/utils.js
+/sdk.js
 
 # Ignore all files in the icons folder
 icons/*
