+++ .npmignore
@@ -12,6 +12,6 @@
 !index.mjs
 !index.d.ts
 !types.d.ts
-!utils.mjs
-!utils.js
-!utils.d.ts
+!sdk.mjs
+!sdk.js
+!sdk.d.ts
