+++ .prettierignore
@@ -13,4 +13,4 @@ scripts/build/templates/*.js
 index.js
 index.mjs
 index.d.ts
-utils.js
+sdk.js
